# Guid: optimize comparison functions